requirements.txt
discord.py[voice]
pydub